Sum to seventh cell
Now lets update "createNewRow" function. In the new rows, leave the fifth and the sixth cell empty.

In the seventh cell put the total sum of the values from seventh cell in the rows that have the same value both in the first and the third cell, as the new row created ad "createNewRow" function.